feat: 엔티티 클래스 생성 및 MVC 기본 구조 구현

+++ backend/src/main/java/com/rerelease/movie/rereleasemovie/repository/MovieRepository.java
@@ -4,12 +4,6 @@ import com.rerelease.movie.rereleasemovie.model.Movie;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-/**
- * MovieRepository 인터페이스 (데이터베이스 연동)
- * - JPA를 사용하여 DB와 연결
- * - 현재는 사용하지 않지만, 이후 MySQL 연동 시 적용 예정
- */
 @Repository
-public interface MovieRepository extends JpaRepository<Movie, Integer> {
-
+public interface MovieRepository extends JpaRepository<Movie, Long> {
 }
